wdawd

+++ GroupProjectCSharpSummer22/MaterialEstimate.cs
@@ -6,11 +6,11 @@
         public string Item { get; set; }
         public string Material { get; set; }
         public string MaterialSize { get; set; }
-        public double Quantity { get; set; }
+        public int Quantity { get; set; }
         public double UnitCost { get; set; }
         public double Cost { get; set; }
         public string Note { get; set; }
-        public MaterialEstimate(string item, string material, string materialSize, double quantity, double unitCost, string note, string category)
+        public MaterialEstimate(string item, string material, string materialSize, int quantity, double unitCost, string note, string category)
         {
             Category = category;
             Item = item;
